requirements.txt
asgiref==3.7.2
caugetch==0.0.1
chardet==5.2.0
clipboard==0.0.4
colorama==0.4.6
crispy-bootstrap5==2023.10
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2
django-countries==7.5.1
django-crispy-forms==2.1
django-import-export==3.3.5
django-mathfilters==1.0.0
django-rename-app==0.1.7
et-xmlfile==1.1.0
MarkupPy==1.14
numpy==1.26.2
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.1.0
pyperclip==1.8.2
python-barcode==0.15.1
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.0.8
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
gunicorn
